+++ src/Resie.jl
@@ -38,21 +38,8 @@ function run_simulation(project_config::Dict{AbstractString,Any})
         step_order = calculate_order_of_operations(components)
     end
 
-    time_step = 900
-    if "time_step_seconds" in keys(project_config["simulation_parameters"])
-        time_step = UInt(project_config["simulation_parameters"]["time_step_seconds"])
-    end
-
-    start_timestamp = 0
-    if "start" in keys(project_config["simulation_parameters"])
-        start_timestamp = Integer(project_config["simulation_parameters"]["start"])
-    end
-
-    end_timestamp = 900
-    if "end" in keys(project_config["simulation_parameters"])
-        end_timestamp = Integer(project_config["simulation_parameters"]["end"])
-    end
-
+    # get time steps from input file
+    time_step, start_timestamp, end_timestamp = get_timesteps(project_config["simulation_parameters"])  
     nr_of_steps = UInt(max(1, (end_timestamp - start_timestamp) / time_step))
 
     parameters = Dict{String,Any}(
@@ -62,82 +49,25 @@ function run_simulation(project_config::Dict{AbstractString,Any})
     )
     EnergySystems.set_timestep(parameters["time_step_seconds"])
 
-    outputs = output_keys(components, project_config["io_settings"]["output_keys"])
-    reset_file(project_config["io_settings"]["output_file"], outputs)
+    # get list of requested output keys for lineplot and csv export
+    output_keys_lineplot, output_keys_to_csv = get_output_keys(project_config["io_settings"], components)
+    do_create_plot = !(output_keys_lineplot === nothing)
+    do_write_CSV = !(output_keys_to_csv === nothing)
 
-    ### set data for profile line plot
-    # check if profile line plot should be created (checks if output_plot is present or if
-    # it is "nothing")
-    plot_bool = !(
-        !(haskey(project_config["io_settings"], "output_plot"))
-        ||
-        project_config["io_settings"]["output_plot"] === "nothing"
-    )
-
-    if plot_bool
-        # set keys for output plots    
-        outputs_plot_keys = Vector{EnergySystems.OutputKey}()
-        for plot in project_config["io_settings"]["output_plot"]
-            append!(outputs_plot_keys, output_keys(components, plot[2]["key"]))
-        end
-
-        # prepare array for output plots
-        outputs_plot_data = zeros(Float64, nr_of_steps, 1 + length(outputs_plot_keys))
+    # Initialize the array for output plots
+    if do_create_plot
+        output_data_lineplot = zeros(Float64, nr_of_steps, 1 + length(output_keys_lineplot))
     end
-
-    ### prepare array for output of all energy flow of all system interfaces
-    # get number of system interfaces for preallocation and medium, source and target of
-    # each interface for sankey diagram
-    nr_of_interfaces = 0
-    medium_of_interfaces = []
-    output_all_sourcenames = []
-    output_all_targetnames = []
-    for each_component in components
-        for each_outputinterface in each_component[2].output_interfaces
-            if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Touple
-                if isdefined(each_outputinterface[2], :target)
-                    # count interface
-                    nr_of_interfaces += 1
-
-                    #get name of source and sink
-                    push!(output_all_sourcenames, each_outputinterface[2].source.uac)
-                    push!(output_all_targetnames, each_outputinterface[2].target.uac)
-
-                    # get name of medium
-                    if isdefined(each_outputinterface[2].target, :medium)
-                        push!(medium_of_interfaces, each_outputinterface[2].target.medium)
-                    elseif isdefined(each_outputinterface[2].source, :medium)
-                        push!(medium_of_interfaces, each_outputinterface[2].source.medium)
-                    else
-                        push!(medium_of_interfaces, each_outputinterface[1])
-                    end
-                end
-            elseif isdefined(each_outputinterface, :target)
-                # count interface
-                nr_of_interfaces += 1
-
-                #get name of source and sink
-                push!(output_all_sourcenames, each_outputinterface.source.uac)
-                push!(output_all_targetnames, each_outputinterface.target.uac)
-
-                # get name of medium
-                if isdefined(each_outputinterface.target, :medium)
-                    push!(medium_of_interfaces, each_outputinterface.target.medium)
-                elseif isdefined(each_outputinterface.source, :medium)
-                    push!(medium_of_interfaces, each_outputinterface.source.medium)
-                else
-                    println("Warning: The name of the medium was not detected. This may lead to wrong colouring in Sankey plot.")
-                end
-            end
-        end
+    # reset CSV file
+    if do_write_CSV
+        reset_file(project_config["io_settings"]["output_file"], output_keys_to_csv)
     end
-    println(
-        length(medium_of_interfaces) !== nr_of_interfaces
-        ? "Warning: error in extracting information from input file for sankey plot."
-        : ""
-    )
+   
+    # get infomration about all interfaces for Sankey
+    nr_of_interfaces, medium_of_interfaces, output_sourcenames_sankey, output_targetnames_sankey = get_interface_information(components)
+   
     # preallocate for speed: Matrix with data of interfaces in every timestep
-    output_all_values = zeros(Float64, nr_of_steps, nr_of_interfaces)
+    output_interface_values = zeros(Float64, nr_of_steps, nr_of_interfaces)
 
     # export order or operatin (OoO)
     if project_config["io_settings"]["dump_info"]
@@ -160,39 +90,19 @@ function run_simulation(project_config::Dict{AbstractString,Any})
             end
         end
 
-        # output to file
-        write_to_file(
-            project_config["io_settings"]["output_file"],
-            outputs,
-            parameters["time"]
-        )
-
-        # get all data of all interfaces in every timestep for Sankey
-        # if the balance of an interface was not zero, the actual energy that was flowing
-        # is written to the outputs.
-        # Attention: This can lead to overfilling of demands which is currenlty not visible
-        # in the sankey diagram!
-        n = 1
-        for each_component in components
-            for each_outputinterface in each_component[2].output_interfaces
-                if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Touple
-                    if isdefined(each_outputinterface[2], :target)
-                        output_all_values[steps, n] = calculate_energy_flow(each_outputinterface[2])  
-                        n += 1
-                    end
-                elseif isdefined(each_outputinterface, :target)
-                    output_all_values[steps, n] = calculate_energy_flow(each_outputinterface) 
-                    n += 1
-                end
-            end
+        # write requested output data of the components to CSV-file
+        # This is currently done in every time step to keep data even if 
+        # an error occurs.
+        if do_write_CSV
+            write_to_file(project_config["io_settings"]["output_file"], output_keys_to_csv, parameters["time"])
         end
 
-        # gather data for profile line plot (@ToDo: may extract from all data in post processing)
-        if plot_bool
-            outputs_plot_data[steps, :] = geather_output_data(
-                outputs_plot_keys,
-                parameters["time"]
-            )
+        # get the energy transported through each interface in every timestep for Sankey
+        output_interface_values[steps, :] = collect_interface_energies(components, nr_of_interfaces)
+      
+        # gather output data of each component for line plot
+        if do_create_plot
+            output_data_lineplot[steps, :] = geather_output_data(output_keys_lineplot, parameters["time"])
         end
 
         # simulation update
@@ -200,12 +110,12 @@ function run_simulation(project_config::Dict{AbstractString,Any})
     end
 
     ### create profile line plot
-    if plot_bool
-        create_profile_line_plots(outputs_plot_data, outputs_plot_keys, project_config["io_settings"]["output_plot"])
+    if do_create_plot
+        create_profile_line_plots(output_data_lineplot, output_keys_lineplot, project_config)
     end
 
     ### create Sankey diagram
-    create_sankey(output_all_sourcenames, output_all_targetnames, output_all_values, medium_of_interfaces, nr_of_interfaces)
+    create_sankey(output_sourcenames_sankey, output_targetnames_sankey, output_interface_values, medium_of_interfaces, nr_of_interfaces)
 
 end