+++ src/energy_systems/general/storage.jl
@@ -109,7 +109,11 @@ function load(unit::Storage, parameters::Dict{String,Any})
 end
 
 function output_values(unit::Storage)::Vector{String}
-    return ["IN", "OUT", "Load", "Load%", "Capacity"]
+    return [string(unit.medium)*" IN",
+            string(unit.medium)*" OUT",
+            "Load",
+            "Load%",
+            "Capacity"]
 end
 
 function output_value(unit::Storage, key::OutputKey)::Float64