+++ src/file_output.jl
@@ -1,5 +1,162 @@
 # this file contains functionality for writing output of the simulation to files.
 
+
+"""
+get_output_keys(config[io_settings], components)
+
+This function determines both for the lineplot and the csv output if
+they should be created:
+    - if not, "nothing" will be returned as key list
+    - if yes, the key lists of the outputs will be returned, either containing
+        - all possible keys if this is requested in the input file or
+        - only the requested keys as requested in the input file
+"""
+function get_output_keys(io_settings::Dict{String, Any}, 
+                         components::Grouping
+                         )::Tuple{Union{Nothing, Vector{EnergySystems.OutputKey}}, Union{Nothing, Vector{EnergySystems.OutputKey}}}
+    # determine if lineplot and csv should be created
+    do_plot_all_outputs = false
+    do_create_plot = false
+    if haskey(io_settings, "output_plot")
+        do_create_plot = true  # if the key exists, set do_create_plot to true by default
+        if io_settings["output_plot"] == "all"
+            do_plot_all_outputs = true
+        elseif io_settings["output_plot"] == "nothing"
+            do_create_plot = false
+        end
+    end
+
+    do_write_all_CSV_outputs = false
+    do_write_CSV = false
+    if haskey(io_settings, "output_keys")
+        do_write_CSV = true  # if the key exists, set do_create_plot to true by default
+        if io_settings["output_keys"] == "all"
+            do_write_all_CSV_outputs = true
+        elseif io_settings["output_keys"] == "nothing"
+            do_write_CSV = false
+        end
+    end
+
+     # collect all possible outputs of all units if needed
+    if do_plot_all_outputs || do_write_all_CSV_outputs
+        all_output_keys = Vector{EnergySystems.OutputKey}()
+        for unit in components
+            temp_dict = Dict{String, Any}(unit[2].uac => output_values(unit[2]))
+            append!(all_output_keys, output_keys(components, temp_dict))
+        end
+    end
+
+    # collect output keys for lineplot and csv output
+    if do_create_plot  # line plot
+        output_keys_lineplot = do_plot_all_outputs ? all_output_keys : Vector{EnergySystems.OutputKey}()
+        # Collect output keys if not plotting all outputs
+        if !do_plot_all_outputs
+            for plot in io_settings["output_plot"]
+                key = plot[2]["key"]
+                append!(output_keys_lineplot, output_keys(components, key))
+            end
+        end
+    else
+        output_keys_lineplot = nothing
+    end
+
+    if do_write_CSV  # csv export
+        if do_write_all_CSV_outputs # gather all outputs
+            output_keys_to_csv = all_output_keys
+        else  # get only requested output keys from input file for CSV-export
+            output_keys_to_csv = output_keys(components, io_settings["output_keys"])
+        end
+    else
+        output_keys_to_csv = nothing
+    end
+
+    return output_keys_lineplot, output_keys_to_csv
+end
+
+"""
+get_interface_information(components)
+
+
+Function to gather information for the sankey diagram.
+Determines 
+- the total number of present system interfaces [int]
+- the corresponding medium in each interface [medium]
+- the source component of each interface and [unit]
+- the target component of each interface [unit]
+
+The information is returned as single vectors with the indicees matching together.
+"""
+function get_interface_information(components::Grouping)::Tuple{Int64,Vector{Any},Vector{Any},Vector{Any}} 
+    nr_of_interfaces = 0
+    medium_of_interfaces = []
+    output_sourcenames_sankey = []
+    output_targetnames_sankey = []
+    for each_component in components
+        for each_outputinterface in each_component[2].output_interfaces
+            medium = nothing  # reset medium
+            if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Touple
+                medium = each_outputinterface[1] # then, the medium is stored separately
+                each_outputinterface = each_outputinterface[2]
+            end
+
+            if isdefined(each_outputinterface, :target)
+                # count interface
+                nr_of_interfaces += 1
+
+                #get name of source and sink
+                push!(output_sourcenames_sankey, each_outputinterface.source.uac)
+                push!(output_targetnames_sankey, each_outputinterface.target.uac)
+
+                # get name of medium
+                if isdefined(each_outputinterface.target, :medium)
+                    push!(medium_of_interfaces, each_outputinterface.target.medium)
+                elseif isdefined(each_outputinterface.source, :medium)
+                    push!(medium_of_interfaces, each_outputinterface.source.medium)
+                elseif !(medium === nothing)
+                    push!(medium_of_interfaces, medium)
+                else
+                    println("Warning: The name of the medium was not detected. This may lead to wrong colouring in Sankey plot.")
+                end
+            end
+        end
+    end
+    println(
+        length(medium_of_interfaces) !== nr_of_interfaces
+        ? "Warning: error in extracting information from input file for sankey plot."
+        : ""
+    )
+
+    return nr_of_interfaces, medium_of_interfaces, output_sourcenames_sankey, output_targetnames_sankey
+end
+
+"""
+collect_interface_energies(components, nr_of_interfaces)
+
+Collects and returns the energy that was transportet through every interface.
+If the balance of an interface was not zero, the actual energy that was flowing
+is written to the outputs.
+Attention: This can lead to overfilling of demands which is currenlty not visible
+in the sankey diagram!
+"""
+function collect_interface_energies(components::Grouping, nr_of_interfaces::Int)
+    n = 1
+    energies = zeros(Float64, nr_of_interfaces)
+    for each_component in components
+        for each_outputinterface in each_component[2].output_interfaces
+            if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Touple
+                if isdefined(each_outputinterface[2], :target)
+                    energies[n] = calculate_energy_flow(each_outputinterface[2])  
+                    n += 1
+                end
+            elseif isdefined(each_outputinterface, :target)
+                energies[n] = calculate_energy_flow(each_outputinterface) 
+                n += 1
+            end
+        end
+    end
+    return energies
+end    
+
 """
 output_keys(from_config)
 
@@ -144,29 +301,43 @@ create a line plot with data and label. user_input is dict from input file
 function create_profile_line_plots(
     outputs_plot_data::Matrix{Float64},
     outputs_plot_keys::Vector{EnergySystems.OutputKey},
-    user_input::Dict{String,Any}
+    project_config::Dict{AbstractString, Any}
 )
-
-    # set Axis, unit and scale factor
-    axis = String[]
-    unit = String[]
-    scale_fact = Float64[]
-    for plot in user_input
-        push!(axis, string(plot[2]["axis"]))
-        push!(unit, string(plot[2]["unit"]))
-        push!(scale_fact, plot[2]["scale_factor"])
-    end
-
-    # create legend entries
-    labels = String[]
-    n = 1
-    for outkey in outputs_plot_keys
-        if outkey.medium === nothing
-            push!(labels, string("$(outkey.unit.uac) $(outkey.value_key) [$(unit[n])] ($(axis[n]))"))
-        else
-            push!(labels, string("$(outkey.unit.uac) $(outkey.medium) $(outkey.value_key) [$(unit[n])] ($(axis[n]))"))
+    plot_all = project_config["io_settings"]["output_plot"] == "all"
+    
+    # set Axis, unit and scale factor if given
+    if plot_all  # plot all outputs. Here no units or scaling factors are available.
+        labels = String[]
+        n = 1
+        for outkey in outputs_plot_keys
+            if outkey.medium === nothing
+                push!(labels, string("$(outkey.unit.uac) $(outkey.value_key)"))
+            else
+                push!(labels, string("$(outkey.unit.uac) $(outkey.medium) $(outkey.value_key)"))
+            end
+            n += 1
         end
-        n += 1
+    else # plot only defined outputs. Here units and scaling factors are available.
+        axis = String[]
+        unit = String[]
+        scale_fact = Float64[]
+        for plot in project_config["io_settings"]["output_plot"]
+            push!(axis, string(plot[2]["axis"]))
+            push!(unit, string(plot[2]["unit"]))
+            push!(scale_fact, plot[2]["scale_factor"])
+        end
+   
+        # create legend entries
+        labels = String[]
+        n = 1
+        for outkey in outputs_plot_keys
+            if outkey.medium === nothing
+                push!(labels, string("$(outkey.unit.uac) $(outkey.value_key) [$(unit[n])] ($(axis[n]))"))
+            else
+                push!(labels, string("$(outkey.unit.uac) $(outkey.medium) $(outkey.value_key) [$(unit[n])] ($(axis[n]))"))
+            end
+            n += 1
+        end
     end
 
     # create plot
@@ -174,11 +345,15 @@ function create_profile_line_plots(
     y = outputs_plot_data[:, 2:end]
     traces = GenericTrace[]
     for i in axes(y, 2)
-        trace = scatter(x=x / 60 / 60, y=scale_fact[i] * y[:, i], mode="lines", name=labels[i])
-        if axis[i] == "right"
-            trace.yaxis = "y2"
-        else  # default is left axis
-            trace.yaxis = "y1"
+        if plot_all
+            trace = scatter(x=x / 60 / 60, y=y[:, i], mode="lines", name=labels[i])            
+        else
+            trace = scatter(x=x / 60 / 60, y=scale_fact[i] * y[:, i], mode="lines", name=labels[i])
+            if axis[i] == "right"
+                trace.yaxis = "y2"
+            else  # default is left axis
+                trace.yaxis = "y1"
+            end
         end
         push!(traces, trace)
     end