+++ CHANGELOG.md
@@ -14,6 +14,11 @@ versions prior to 1.0.0 any release might break compatability. To alleviate this
 the meaning of major-minor-patch is "downshifted" to zero-major-minor. However some
 breaking changes may slip beneath notice.
 
+## Version 0.6.5
+* "output_keys" and "output_plot" in the input file can now be "all", "nothing" or a list of entries for custom outputs of the CSV file and lineplot (backwards compatibility is given)
+* output_values() of all components was changed to not only return the available channels, but also the corresponding media
+* restructured "run_simulation()" for better readability
+
 ## Version 0.6.4
 * Generalise implementation of gas boiler to that of a fuel boiler
   * Although this implicates the use of a chemical fuel to generate heat, the current implementation works with any kind of input including electricity